function/requirements.txt
azure-functions==1.7.0
psycopg2-binary==2.8.6
python-http-client==3.3.2
sendgrid==6.7.0
starkbank-ecdsa==1.1.0
